+++ resources/views/components/browser/row.blade.php
@@ -17,9 +17,11 @@
     aria-selected="{{ $isSelected ? 'true' : 'false' }}"
     data-name="{{ $item->name }}"
     data-kind="{{ $item->kind }}"
+    data-directory="{{ $item->directory }}"
     data-url="{{ $item->publicUrl() }}"
     x-on:click="pick({{ $loop->index }}, {!! $js($item->id) !!}, $event)"
     x-on:dblclick.prevent="openPreviewAt({{ $loop->index }})"
+    x-on:contextmenu.prevent="openContextMenu({{ $loop->index }}, $event)"
 >
     <td class="fml-table__check" x-on:click.stop>
         <x-filament::input.checkbox